test(golden): golden file tests for ctrls (#1386)
This adds the helper function `compareWithGolden` to the controller
package. This function can be used to compare the output of a
controller's action (e.g. Show, Create, List, etc.) against a file that
is considered to be the truth (the golden file).

`compareWithGolden` takes a test, the path to a golden file and an object
as its arguments. The object can be anything basically but in the
controller package it is the API response, so either the result of a
successful API call (e.g. a JSONAPI object) or a JSONAPI error.

`compareWithGolden` converts the actual object to a JSON string in order
to allow for deep comparison of nested values. It then loads the golden
file and compares it against the JSON. If the two are different, the
given test fails and a nice diff message is printed to the screen.

Golden files are not supposed to be written by hand. There's an easy
mechanism to create a first version of them by specifying an `-update`
flag that is only available in `go test`.

Here's an example for how to create a golden file for the only test that
currently uses this approach:

```sh
ALMIGHTY_POSTGRES_HOST=172.17.0.2 \
ALMIGHTY_RESOURCE_DATABASE=1 \
go test -v github.com/almighty/almighty-core/controller \
  -run TestRunCodebasesTest \
  -testify.m TestSuccessShowCodebaseWithoutAuth \
  -update
```

This will run the test `TestSuccessShowCodebaseWithoutAuth` as normal
but it will take the actual object passed to `compareWithGolden` as the
new truth. The golden file is then updated.

The benefit of this approach of testing is three fold:

1. We have examples of how our response looks like. This is somewhat
self-documenting.

2. Previously when we checked that an error was returned from an action
we only did check that an error occured but we didn't pay attention to
the real error message. Now, `compareWithGolden` takes care of this for
us.

3. When you want to compare a value inside of a JSONAPI relationship in
Go code you first need to check all the pointers to that deeply nested
value. Now we can compare everything using just strings.

One drawback is that you have to make sure that IDs in the golden file
don't change from each execution of your test. That's why I've
introduced the sticky IDs for the space and codebase inside of the go
code.
Currently we cannot use golden files for spaces because the owner ID
changes between each test execution. I think PR #1359 fixes this issue
(at least according to @alexeykazakov).

Another drawback is that you need to turn off Gorm's automatic time
value generation for `created_at` and `updated_at` values. I've done
this explicitly for this test.