Use ctx in newinternalerror (#1409)
This PR adds a `context.Context` parameter to the `errors.NewInternalError()` function. With this change we have everything in place to go ahead and log errors whenever an `errors.InternalError` is created. The idea is to have a centralized place for logging errors of this kind instead of having to log the error manually before it is returned from a function.

This PR doesn't contain the logging in `errors.NewInternalError` because there's some stupid cyclic import dependency if we include our `log` package in the `errors` package. Adding the log call can be part of a future PR.