Adding CheckExists that only returns an error
Adding this convenience function if you're only interesting in the existence of an element and maybe want to pass any error through.

This avoids having to write:

```go
// The more inconvenient way to check for existence:
if _, err := repo.Exists(ctc, "foo", id): err != nil { /*...*/ }

// Adding a little more convenience
if err := repo.CheckExists(ctc, "foo", id): err != nil { /*...*/ }
```

The `Exists` function itself always returns a not-nil error value if the object with the given id doesn't exist. That's why we can focus on the error and leave the boolean return value alone.