Record timestamp of changes related to work items  (#1382)
Added triggers at the DB level to fill the new `relationships_changed_at`
column when a comment is added or removed on
a work item, or when a work item is linked/unlinked to another
one (in which case, both source and target items are affected
by the trigger)
Using a private field in 'workitem.Workitem' to store the timestamp

Also, renamed some utility method to reflect their purpose
(ie, return a payload for entity creation request, not actually create an entity)

Also fixed a bug where the `created_at` column would be reset
if the payload did not contain the previous `created_at` value
(which is not mandatory in the API). The data is now retrieved
from the actual record before the SQL update is performed.

Use the new `relationshipsChangedAt` field to compute the `ETag`
and `Last-Modified` values and perform the comparisons when processing
a request.

Removing the test utility function that computes an ETag from work items
since the new `relationshipsChangedAt` field is private. Instead,
we need to query the server and use the `ETag` response header in
subsequent requests.

Fixes #1382 and fixes #1370

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>