Handle mapping of field names to SQL columns/jsonb
Alongside the old "ID", "Type", and "Version" fields I've added the
"Number" and "SpaceID" fields.

A little change was necessary to add that little "_" when converting the
"SpaceID" to the "space_id" field.

Now, postgres might not differentiate between "ID" and "id" when it
comes to column names, but I figured that it won't do any harm if I add
an explicit mapping of struct field names to SQL columns. I needed it
for the "SpaceID" struct field only but I've added it for the rest of
the top-level columns (non-JSONB fields) as well.