Etag for iterations does not change when number of attached WIs changed (#1587)
Added a `relationships_changed_at` column in the `iterations` table to record
the timestamp of changes in work items when they are linked/unlinked to an
iterations, or soft-deleted.
The filling of the `relationships_changed_at` is performed by a
set of triggers and functions.

Added tests to verify that the `ETag` and `Last-Modified` response headers
change when a work item is added to/removed from an iteration.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>