Added test fixture package (#1612)
## Synopsis 

This change introduces the concept of a so called **test fixture**. A test fixture when created is given a list of **recipe functions** of things to create. After creation, the test fixture contains arrays with created objects for tests to use.

## Purpose

The purpose of this PR is to have a framework that simplifies test setup. Right now every test (be it on repo or controller level) needs to manually create and identity, a space and all the other dependencies just in order to get to a state in which the test can run the actual action under test.

Every time we change some of the dependencies, all tests need to adapt and that caused problems in the past.  With the new test fixture, a recipe function like `Spaces(10)` **implicitly triggers** (if not told differently) the creation of an identity and then creates 10 spaces. It's best if you look at the provided tests in `testfixture_test.go` to see the power of this concept including manual intervention with **customize-entitiy-callbacks**. Just an example, can you imagine the effort to create 100 comments from 100 different users on a single work item? Then now see how simple it is with a test fixture:

```go
ctx := NewTestFixture(s.T(), s.DB, Identities(100), Comments(100, func(fxt *TestFixture, idx int) error {
	fxt.Comments[idx].Creator = fxt.Identities[idx].ID
	return nil
}))
```

## Benefit 

If a test is only interested in testing something with comments, it should not need to be forced to deal with identity, space, work item type, and work item creation. Now it doesn't have to know but if needed, the user can always intercept the creation and do complicated distribution calculations or something like that.