Redirect GET space/collaborators to Auth service (#1657)
Requests to fetch collaborators' information from WIT ( api.openshift.io ) will be redirected to AUTH ( auth.openshift.io ). This also ensures that WIT doesn't interact with Keycloak directly anymore. 

For ease of development in local setups,

In development mode with F8_AUTHZ_ENABLED=false, the remote call to Auth is skipped, and only the space owner info is returned as the lone collaborator.

In development mode with F8_AUTHZ_ENABLED=true, the remote call to Auth is made.