Refactor the `SearchByText` test
Refactored the test to use the TestFixture to initialise the
work items from the data set.
Remove the `optionalKeywords` which contained the work item number
but failed when the 3rd work item was retrieved because its number
('3') would match the description. This worked in the past when the
number did not exist and the search was based on the ID (UUID).
Also, searching on a URL did not seem to make sense since there was
no such value in the work item fields.

Also, add a test to search by number using the `number:%d` query.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>