representation of "empty" labels & assignees
When a new work item is created without lables, then labels data is
stored as "NULL" in the database. When the same work item is update
without labels, it store "[]" as the value in the database. This
behaviour is same for assignees as well.

If the value is empty, the entire key/value pair will be removed.

Fixes https://openshift.io/openshiftio/openshiftio/plan/detail/1517