Improved searchRepositoryWhiteboxTest.TestSearchByID
Before we only tested the search by WI number without specifying a space. This was problematic because there can be many WIs with the same number in different spaces. This is because a WI number is only unique within a space.

Now we test four rather than one case:

1. Test that a workitem can be found only its number and that it is
among the list of found WIs.

2. Test that we cannot find a WI by its number by searching for a number
that doesn't exist.

3. Same as 1. but specify a space as well.

4. Same as 2. but specify a space as well.