Add an endpoint on WIT to search for codebases by URL and provide the space owner (openshiftio/openshift.io#1246)
Added an endpoint that allows for searching codebases by their URL, with support
for pagination (in case we would have multiple codebases pointing to the same queried
URL)
Also, response includes the data of the relationship spaces, so the consumer can determine
the ID of the owner of each matching codebase
This also requires 2 additional methods in the repositories, to search for codebases by
URL and list a group spaces by their IDs
Also, rename the `Test` function and `TestSuite` struct in the `space/space_test.go` file
to better match with the actual content.
Also, refactor the way Spaces are converted to JSON by moving the backlog/totalCount into
a `SpaceConvertFunc` so it can be optional.

Fixes openshiftio/openshift.io#1246

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>