Fix test failures with golden file ((openshiftio/openshift.io#1246))
Use a different ordering to make sure that the test pass no matter which
ID has been allocated to the codebases and their spaces.
Ordering the codebases in `data` by the ID of their related space, and
ordering the `included` spaces by their ID ensures that all data are in
such an order, than replacing all ID with incremental UUIDs in the files
still works.

Fixes openshiftio/openshift.io#1246

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>