UUID agnostic golden file comparison replaces real UUIDs during update (#1774)
When you update a golden file using `go test -update` it will contain a set of new random UUIDs.

Those new UUIDs appear as changes inside a commit if you just do `git add myGoldenFile.json`. This clutters the commit and hides the real change that might have appeared in your commit.

I want you to be able to focus on the real changes to the API and not on some random UUIDs.

That is why a golden file that is generated with `controller.compareWithGoldenUUIDAgnostic()` will now have all UUIDs replaced with  "00000000-0000-0000-0000-000000000001", "00000000-0000-0000-0000-000000000002", ..., "00000000-0000-0000-0000-00000000000N" and so on.

The next time you update your golden files will be much easier.