Add next/prev group relationship to wit group (#1797)
In order to help a client navigate around the work item type groups I've added a relationship to the previous and next work item type group (if any). When only looking at one work item type group (not at the list), this should help to get the next or previous one. Also this should help when adding a guided quick add.