WIP: Improve API doc for iteration (#1806)
## Problem

Based on a discussion between Ruchir and me, we've discovered that our core API is currently hard to use because we're missing real world examples and a good swagger documentation.

The most important part to get fixes is to tell the user which fields in a resource payload are **MANDATORY** for what kind of action.

For example, there's no documentation on what attributes or fields are required when creating or updating an iteration.

## Proposal

### Change generated swagger documentation

In order to address this, I've created a way to mark the description of fields to be mandatory.

Before this was the description for the the `name` attribute of an iteration.

```go
a.Attribute("name", d.String, "The iteration name", nameValidationFunction)
```

You have to understand that we re-use the whole iteration construct during all of the supported actions (e.g. create, update, delete, etc.). That means we cannot simply mark it as mandatory once and for all. But instead we can do something like this:

```go
a.Attribute("name", d.String, mandatoryOnCreate("The iteration name"), nameValidationFunction)
```

### Real world examples

Instead of maintaining a set of real world example payloads that might run out of date we might as well just dump test requests.

In our tests we used to only have golden files for responses but not for the actual requests. Now I've added the requests as well. Just by looking at this file for example, it should be possible to find out what attributes are required to create an iteration:

**controller/test-files/iteration/create/ok_create_child_ID_paylod.req.payload.golden**:
```json
{
  "data": {
    "attributes": {
      "description": "Some description",
      "endAt": "2016-11-25T15:08:41Z",
      "name": "Sprint #21",
      "startAt": "2016-11-04T15:08:41Z",
      "user_active": false
    },
    "id": "00000000-0000-0000-0000-000000000001",
    "type": "iterations"
  }
}
```

Now, this might not be the bare minimum you need to specify but if you now take a quick look at the new swagger portion, you can find out really easy what is needed.

http://swagger.goa.design/?url=github.com%2Ffabric8-services%2Ffabric8-wit%2Fdesign#!/iteration/iteration_create_child