Separate Guided Child creation and Browsing (#1820)
The execution group shall also contain the Bug and the Feature WIT.
That effectively means that a WIT does appear in more than one group
now. The group should now only be used to do the viewing part and not
the guided child creation part.

For the guided child creation I have modified the work item type
response to include a new relationship called `guidedChildTypes`.
We must this relationship when proposing child types to the user.

Currently this is all hardcoded and only available for the
system-defined work item types in order to get going UI-wise.