Link updates (#1822)
Removes **Update**/**Save** method of work item link repository and all controllers.

The only permitted actions to a link once created are:
 * **delete**
 * **show**
 * **list**

In order to mimic the **update** behavior you have to **delete and recreate** a link with different settings (e.g. with a different link type). This is needed in order for us to maintain a valid *topology*.

Also the `Create` method has been removed from the `/api/workitems/<ID>/relationships/links` endpoint. Only the `List` method remains for this endpoint.