introduce "ptr" package for on-the-fly pointer values (#1830)
Package `ptr` provides functions to create on the fly pointer values for some built-in types. This might seem stupid at first but in our codebase we have a lot of places in which we just need to create variable just to get the pointer. That clutters the code a lot and causes distraction. Most of the time those strings are just throw-away constants and the pointer "is not really important later".

I've also changed the function `newCreateChildAreaPayload` to accept a `string` rather than a `*string` because in 99% of the time the string is not nil.