add parameter to sort the output
You can provide parameter to sort the output. The values that can be
given to the parameter are following:

"name", "owner", "created", "updated",
"-name", "-owner", "-created", "-updated".

The parameter values which are appended with '-' sign will sort the
output in descending order of the respective parameter, rest will
sort in ascending order.

The 'sort' parameter is available to the API endpoint
`/api/search/spaces` and `/api/namedspaces/:username`.