Space templates
* Work items, link types and spaces now point to a space template
instead of a space
* A space template in the DB consists of nothing more than an ID, a name
and a description.
* A space template in the broader sense describes a combination of work
item types, link types and type groups in a YAML file that is imported
at core startup. I'm using gopkg.in/yaml.v2 for the handling of YAML
files.
* Controllers for space templates supports "show" and "list" actions.
The latter is located under the "spacetemplates" endpoint but the first
is the standalone "workitemtypes" endpoint. The same is true for work
item link types.
* Modified conditional request generator in order to support caching
for space templates.
* Database migration changes for space templates.
* Test fixture updated to support space templates.
* Fixed enum base type from integer to string.

Currently the core has no way of checking that all provided enum values
actually are of a certain type and there's no error because every value
is stored in an interface{} object.

There's a note in workitem.ConvertToModel():

"// to be done yet | not sure what to write here as of now."

which marks the place to add such code in the future.

* space template name must not start with _ or - and has to be 1 to 62
characters (see `design` folder).

* Added space relationship as obsolete
* Added plantuml class diagram for space templates and related stuff.
This is only for educational purposes and not meant to be an exact
representation of the real circumstances.
* I've persisted work item type groups in the database and created a
dynamic endpoint for it.