Replace $WITGROUP with typegroup.name (#1929)
`typegroup.name` instead of `$WITGROUP` will be the new field to query for typegroups and the special handling for wit groups in the `search_repository.go` will go away soon. But in order to give the UI a chance to replace `$WITGROUP` with `typegroup.name` I wanted to get this change in as early as possible.