fix(search): support old "$WITGROUP" as well as new "typegroup.name" fields (#1932)
This addresses the backward compatibility problems I introduced with #1929 (was reverted meanwhile).

`typegroup.name` instead of `$WITGROUP` will be the new field to query for typegroups and the special handling for wit groups in the `search_repository.go` will go away soon.

In #1929 I removed support for `$WITGROUP` but here I kept it an only added support for an additional `typegroup.name`. I've modified all tests that deal with `$WITGROUP` to now test for `typegroup.name` as well.