Added DefaultValue function to FieldType interface
For enum types the default value used is the first one of the possible
values.

Fot the simple and list type the DefaultValue will return whatever is
given to the functions. The plan is to specify a default value in the
type definition of the simple and the list field.