Add DefaultValue() to workitem.FieldType (#1992)
The enum type is the only field type that returns a default value if `nil` is given. It returns the first value from the list of allowed values. All other field types return `nil` if `nil` is given. This may change in the future.

Also, the enum type now has a field `RewritableValues` that lets you decide if you want to allow for another work item to overwrite the allowed values for that field. This is needed for the space template branch to allow for each work item type to define it's own states under the same field name `system.state` that was defined before by the planner item type.

* Add DefaultValue() to workitem.FieldType
* Properly initialize workitem.EnumType
* Add unit-tests for workitem.EnumType.DefaultValue()

Outsorced from #1148