fix system.number field storage and add read-only fields (#1996)
## Add read-only attribute to fields

This change adds a new field attribute called `ReadOnly` and marks these fields as read-only:

1. `system.created_at`,
2. `system.updated_at`, 
3. `system.number`, and
4. `system.order`.

The work item repository's `Save` method no longer updates fields that are read-only, namely all of the above).

**NOTE:** As discussed with @DhritiShikhar and @aslakknutsen in the chat today (see [here](https://chat.openshift.io/developers/pl/pz1md7zx83f9trckky3x8sdo7c) and [here](https://chat.openshift.io/developers/pl/3wfrzpx5uidx9e3pj97e3r1tcr)) we agreed to ignore changes to `system.order` that happen with the repo's `Save()` method. For now you can only update a work items execution oder by using the `Reorder()` functions.

### Not related addition of `system.number`

I've noticed that the test fixture make function for work items didn't include the `system.number` field. So I've added that while I was modifying the file because of the `ReadOnly` addition.

## Remove `system.number` from fields

The `system.number` field accidentally became a persisted field when I've added it to the work item description of the planner item type (and thereby to any other WIT). Normally we returned the `system.number` field as a kind of virtual field that is generated based on the value of the `number` column of the work item.

With this change we remove all persisted `system.number` entries from all work items and make sure it won't be stored (see read-only above) as a field again (see #1961).

## Implications for the UI

The UI should render read-only fields as non-editable (@sanbornsen @michaelkleinhenz).

Partially outsourced from #1148