Rework forward and reverse link names (#2039)
This changes the reverse name of the "Blocker" link type from `blocked by` to `is blocked by`. 

Also the forward and reverse name of the "Parenting" link type are prefixed with an `is` now: `is parent of`, `is child of`. 

This is done to have a better experience when creating links manually with the UI where the UI says something like `This work item foo parent of ..." and it would be much nicer if it instead said `This work item foo is parent of`.