Add update-client Make target
The `update-client` make target builds the fabric8-wit client and
compares it with the fabric8-wit-client respository. If there are
updates, there will be a tmp/client-repo/ that has a new commit with all
the changes that went into the upstream project. The client lib is only
updated when there were changes to the design code in the upstream
project.

There's no automation yet to push the updated lib back to github but it
can easily be integrated.

See #2063