Add capacity flag to namespaces - `cluster-capacity-exhausted`
Now when calling `/user/services`, the namespaces information that is
returned by wit service will also include an attribute called
`cluster-capacity-exhausted` which is a boolean indicating if the
cluster this namespace is hosted on has exhausted it's resources or is
still capable of running more workloads.

This information can be used by OSIO services to make decision whether
to spawn any new OSIO service in user namespace on that cluster.

Related to openshiftio/openshift.io#3320