Call WI's conversion from model to to JSONAPI and back using the WI's type (#2069)
In this PR I've changed the signature of the `controller/ConvertWorkItem` and `controller/ConvertWorkItems` function. They now accept the work item type of the work item that is supposed to be converted. The type itself is **not yet** used but it will in a future PR.

This is needed to support generic fields (see #2070).