Validate space name with a regexp in the controller (OSIO#3580)
This will prevent creating spaces with a name that cannot be used
as a value in pod labels (matching max length and pattern).
The regexp and length checks in the controller replace the validation function
that was previously used at the design level, which allows for returning
proper JSON-API errors to the clients.

Fixes openshiftio/openshift.io#3580

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>