Prevent starting old cores on newer databases
When a core starts up and find out that the database version is newer
than the one it is aware of it stops the migration.

NOTE: A core that is already running will not error out if the database
is newer than what it is aware of. This only happens when a core starts
up and the migration happens.

See openshiftio/openshift.io#3880