Fix search for work items by board ID
A table join is now not only validated by the actual columns a table has
but by the columns that are explicitly allowed by a user. This allowed
us to have the board columns joined to each work item as a virtual
table in which we re-defined the "id" column to not be the ID of the
column but the ID of the board itself.