Integrate call to analytics for cve-scanning
Earlier the attribute was added to codebases named `cve-scan` which
determines if this codebase should be scanned or not.

Now this commit makes sure that if `cve-scan` is enabled then call
analytics gemini service to enable this scan.

Everytime a codebase is added it is registered with the analytics
service and when a codebase is deleted or updated a check is made to
make sure that no entry for such a codebase exists and then call to
unsubsribe the codebase from scanning.