send revisionid to notification service
Notification service is only sent the workitem id when it is updated.
This makes the information very limited to generate useful notification.
This commit makes it possible to send the revision id along with
workitem id.

To send the revision id it makes use of the custom attribute of the
notification service. This way notification service does not need to
make any changes to accept this new information.