Codebase can search git@github... type codebase URLs
All the URLs that have git@github.com:USERNAME/REPOSITORY.git
cannot be searched on wit because the URLs added to wit are of the
format https://github.com/USERNAME/REPOSITORY.git

While at build time the URL that is sent to the analytics service is of
the format 'git@github' and when this is plainly queried nothing is
found.

So this commit converts the URL of the format 'git@github...' to the
'https://github...', before search.