Handle area relationships
Moved links to an area/iteration/user/board-column relationship out
of the `data` section in the relationship and into their own `links`
section alongside the `data` section.

Removed links as a whole from area/iteration/label/user/board-column
relationships, because a link as we have it can only link to an
individual entitiy and not to all entities that are part fo the
relationship. For this to work we would need to have support for
links to the relationship of a resource, so for example:

* /api/workitems/1234/author
* /api/workitems/1234/relationship/author

But at least this change fixes some of the problems with links appearing
in the wrong place.