Move links of a relationship outside of the data-section (#2205)
Moved links to an 
* area
* iteration
* user
* board-column

 relationship out of the `data` section in the relationship and into their own `links` section alongside the `data` section.

Removed links as a whole from area/iteration/label/user/board-column relationship**s** (plural!), because a link as we have it can only link to an individual entity and not to all entities that are part for the relationship. For this to work we would need to have support for links to the relationship of a resource, so for example:

* /api/workitems/1234/author
* /api/workitems/1234/relationship/author

This change fixes some of the problems with links appearing in the wrong place. This is to make the API more JSONAPI compliant. It is not the final JSONAPI compliance I'm striving for but a reviewable step in that direction. 

## Background

JSONAPI doesn't allow for 'links' in the `data` section of a relationship as in this example:

```yaml
{
  "data": {
    "type": "articles",
    "id": "1",
    "relationships": {
      "author": {
        "data": {
          "type": "people",
          "id": "9",
          "links": {
            "self": "http://example.com/articles/1/relationships/author",
            "related": "http://example.com/articles/1/author"
          }
        }
      }
    }
  }
}
```

But moving the `links` section next to the `data` makes it a valid JSONAPI response:

```yaml
{
  "data": {
    "type": "articles",
    "id": "1",
    "relationships": {
      "author": {
        "links": {
          "self": "http://example.com/articles/1/relationships/author",
          "related": "http://example.com/articles/1/author"
        },
        "data": {
          "type": "people",
          "id": "9"
        }
      }
    }
  }
}
```




You can try out the above examples here: https://www.jsonschemavalidator.net/

We did include a links in the `data` section in many places.