Updating states on Agile template, relaxing importer (#2170)
This adds some states to the list of available states on Themes and Epics in the Agile template. It also relaxes the template importer so adding new values to an existing enum is now allowed when editing templates. Note that removing values is still not possible without migration and further changes on the code.