Serve swagger under /api/swagger.json and swagger-ui (#2217)
This serves the generated `swagger.json` under `/api/swagger.json` and replaces the host with the current setting from `config.GetHTTPAddress()`. This should make the swagger output compatible with whatever host it is running on (local, prod-preview, prod, ...).

I've also added a swagger-ui container that starts up on `make dev` and is served at [http://localhost:8081/](http://localhost:8081/):

See also #165 for an old version of this.