Replace swagger host with request host and adjust scheme (#2220)
As a follow up for #2217 this change will properly replace the host (currently always `host: "0.0.0.0:8080",`) in the [generated swagger file](https://api.prod-preview.openshift.io/api/swagger.json) with the host of the request that tries to access the swagger file. We strip off the `http://` or `https://` prefix because that's handled in the swagger `schemes` section.

We also set the scheme to `https` if the current request was made using `https://`.