Added cico_run_e2e_tests.sh (#2235)
We want to have end to end tests checking our PRs and in https://github.com/fabric8-services/fabric8-wit/pull/2197 @rgarg1 is working on how this.

It makes sense to run the end to end tests as a standalone job because the tests can be slow and can output messages that we don't want to see inside of our unit or integration tests.

This change adds a `cico_run_e2e_tests.sh` file that fails all the time and outputs a message. Once the file https://github.com/openshiftio/openshiftio-cico-jobs/blob/master/devtools-ci-index.yaml is adjusted to pick up run the end to end tests using this file, then @rgarg1 can adjust his PR to execute the tests in this `.sh` file.