Introduce RevisionID field in work item event (#2253)
Before this change a work item event had an ID that was not unique among events; instead it was the ID of the work item revision in which potentially more than one field can be changed. Since each event deals only with one field change at a time, there potentially were more than one event with the same ID.

This change is supposed to help get better email notifications that @alexeykazakov requested yesterday.