testfixture.GetTestFileAndFunc() (#2149)
`GetTestFileAndFunc` returns the file and function of the first `_test.go` file to appear in the stack and returns it in this schema:

```
(see function github.com/fabric8-services/fabric8-wit/test/testfixture_test.TestGetGetTestFileAndFunc in test/testfixture/testfixture_test.go)
```

The result can be used to augment entities so that we always can tell which test created an entity that is a left-over and not cleaned up for example.

NOTE: The `testfixutre.GetTestFileAndFunc()` was taken out of this PR and already merge with this PR #2214. It is no longer part of this PR.

Most of the entities created with the `test/testfixture` package now contain a description that reveals in which function the entities was created. When there are left-over entities that hinder tests from running properly, you can now go in the database and find out which test produced these left-overs. To simplify things when running on CI, we now output a list of remaining spaces automatically that looks like this if everything went ok:

```
Check that no unexpected left-over spaces exist after running tests:

-[ RECORD 1 ]-------------------------------------

id          | 2e0698d8-753e-4cef-bb7c-f027634824a2

name        | system.space

description | Description of the space
```

Note, that the `system.space` is okay to appear in this list. It is expected even though it should go away at some point.

See https://github.com/fabric8-services/fabric8-wit/pull/2278 for an example of a PR that was easily fixed because of this PR.